+++ cmd/root.go
@@ -1,15 +1,19 @@
 package cmd
 
 import (
+	"os"
+	"path/filepath"
+
 	"github.com/crazy-max/firefox-history-merger/sqlite"
 	. "github.com/crazy-max/firefox-history-merger/utils"
 	"github.com/spf13/cobra"
 )
 
 var (
-	debugEnabled bool
-	placesDb     sqlite.PlacesDb
-	faviconsDb   sqlite.FaviconsDb
+	debugEnabled     bool
+	appPath          string
+	masterPlacesDb   sqlite.PlacesDb
+	masterFaviconsDb sqlite.FaviconsDb
 )
 
 var RootCmd = &cobra.Command{
@@ -20,6 +24,7 @@ More info on ` + AppUrl,
 }
 
 func init() {
+	appPath, _ = filepath.Abs(filepath.Dir(os.Args[0]))
 	cobra.OnInitialize(initRoot)
 	RootCmd.PersistentFlags().BoolVarP(&debugEnabled, "debug", "x", false, "Debug")
 }