+++ sqlite/favicons/moz_icons_to_pages.go
@@ -1,7 +1,7 @@
 package favicons
 
 /**
- * moz_icons_to_pages schema v39
+ * moz_icons_to_pages schema >= v39
  */
 type MozIconsToPages struct {
 	PageId int `gorm:"primary_key"`