+++ utils/path.go
@@ -1,0 +1,13 @@
+package utils
+
+import "strings"
+
+// PathJoin to join paths
+func PathJoin(elem ...string) string {
+	for i, e := range elem {
+		if e != "" {
+			return strings.Join(elem[i:], `\`)
+		}
+	}
+	return ""
+}